+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.89] - 2024-06-17
+### Added
+* `GetMouseWorldPoint` to `ScreenUtils`
+* `Enums` directory and `enum` `Axis`
+### Fixed
+* `GizmosExtensions.DrawBoxCast` where overloaded function did not include `maxDistance`
+
 ## [0.0.88] - 2024-06-12
 ### Added
 * `ColorExtensions.ToNegative` to flip a color